requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg[binary]
python-dateutil
python-multipart
passlib[bcrypt]
python-jose